pyproject.toml
[tool.poetry]
name = "customer-support-assistant"
version = "0.1.0"
description = "Backend for a customer support assistant with AI integration"
authors = ["U7ama <u7amajutt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
aiohttp = "^3.8.6"
groq = "^0.4.0"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"